# swagger/ResolveEnigmaAp - Copy/src/main/java/co/com/bancolombia/_microservice/resolveEnigmaApi/api/GetStepApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.4).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package co.com.bancolombia._microservice.resolveEnigmaApi.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import co.com.bancolombia._microservice.resolveEnigmaApi.model.JsonApiBodyRequest;
import co.com.bancolombia._microservice.resolveEnigmaApi.model.JsonApiBodyResponseError;
import co.com.bancolombia._microservice.resolveEnigmaApi.model.JsonApiBodyResponseSucess;
import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2019-07-17T16:23:36.474-05:00[America/Bogota]")
@Api(value = "getStep", description = "the getStep API")
public interface GetStepApi {

    @ApiOperation(value = "Get one enigma step API", nickname = "getStep", notes = "", response = JsonApiBodyResponseSucess.class, tags={  })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "search results matching criteria", response = JsonApiBodyResponseSucess.class),
        @ApiResponse(code = 424, message = "bad input parameter", response = JsonApiBodyResponseError.class) })
    @RequestMapping(value = "/getStep",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<?> getStep(@ApiParam(value = "Get one enigma step API" ,required=true )  @Valid @RequestBody JsonApiBodyRequest body);

}